# target/classes/static/spec/swagger-v1-desafio-backend.json
{
  "swagger": "2.0",
  "info": {
    "title": "Desafio backend",
    "description": "API para controle de funcionarios e projetos",
    "version": "1.0.0"
  },
  "basePath": "/desafio-backend/v1",
...
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "paths": {
...
    "/projetos": {
      "post": {
        "summary": "Criar projeto",
        "operationId": "createProjeto",
        "description": "Criar projeto",
        "tags": [
          "Projeto"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Novo Projeto",
            "description": "Informações do novo projeto para inclusão.",
            "schema": {
              "$ref": "#/definitions/Projeto"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "$ref": "#/responses/400RequisicaoMalFormada"
          },
          "401": {
            "$ref": "#/responses/401NaoAutorizado"
          },
          "403": {
            "$ref": "#/responses/403NaoPermitido"
          },
          "500": {
            "$ref": "#/responses/500ErroInternoServidor"
          },
          "501": {
            "$ref": "#/responses/501NaoImplementado"
          }
        }
      },
      "get": {
        "summary": "Buscar todos projetos cadastrados",
        "operationId": "searchProjeto",
        "description": "Buscar todos os projetos cadastrados.",
        "tags": [
          "Projeto"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Projeto"
              }
            }
          },
          "400": {
            "$ref": "#/responses/400RequisicaoMalFormada"
          },
          "401": {
            "$ref": "#/responses/401NaoAutorizado"
          },
          "403": {
            "$ref": "#/responses/403NaoPermitido"
          },
          "500": {
            "$ref": "#/responses/500ErroInternoServidor"
          },
          "501": {
            "$ref": "#/responses/501NaoImplementado"
          }
        }
      }
    },
    "/projetos/{id}": {
      "get": {
        "summary": "Buscar projeto pelo identificador",
        "operationId": "searchProjetoById",
        "description": "Buscar projeto pelo identificador.",
        "tags": [
          "Projeto"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do projeto",
            "type": "integer",
            "format": "long",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "schema": {
              "$ref": "#/definitions/Projeto"
            }
          },
          "400": {
            "$ref": "#/responses/400RequisicaoMalFormada"
          },
          "401": {
            "$ref": "#/responses/401NaoAutorizado"
          },
          "403": {
            "$ref": "#/responses/403NaoPermitido"
          },
          "404": {
            "$ref": "#/responses/404NaoEncontrado"
          },
          "500": {
            "$ref": "#/responses/500ErroInternoServidor"
          },
          "501": {
            "$ref": "#/responses/501NaoImplementado"
          }
        }
      },
      "put": {
        "summary": "Atualizar projeto",
        "operationId": "updateProjeto",
        "description": "Atualizar projeto",
        "tags": [
          "Projeto"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do projeto",
            "type": "integer",
            "format": "long",
            "required": true
          },
          {
            "in": "body",
            "name": "Projeto",
            "description": "Informações do projeto para atualização.",
            "schema": {
              "$ref": "#/definitions/Projeto"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "schema": {
              "$ref": "#/definitions/Projeto"
            }
          },
          "400": {
            "$ref": "#/responses/400RequisicaoMalFormada"
          },
          "401": {
            "$ref": "#/responses/401NaoAutorizado"
          },
          "403": {
            "$ref": "#/responses/403NaoPermitido"
          },
          "404": {
            "$ref": "#/responses/404NaoEncontrado"
          },
          "500": {
            "$ref": "#/responses/500ErroInternoServidor"
          },
          "501": {
            "$ref": "#/responses/501NaoImplementado"
          }
        }
      }
    },
...
  },
  "responses": {
...
    "400RequisicaoMalFormada": {
      "description": "Algum parâmetro informado está inválido",
      "schema": {
        "allOf": [
          {
            "$ref": "#/definitions/ApiError"
          },
          {
            "example": {
              "httpStatus": "BAD_REQUEST"
            }
          }
        ]
      }
    },
    "401NaoAutorizado": {
      "description": "Não autorizado",
      "schema": {
        "allOf": [
          {
            "$ref": "#/definitions/ApiError"
          },
          {
            "example": {
              "httpStatus": "UNAUTHORIZED"
            }
          }
        ]
      }
    },
    "403NaoPermitido": {
      "description": "Não permitido",
      "schema": {
        "allOf": [
          {
            "$ref": "#/definitions/ApiError"
          },
          {
            "example": {
              "httpStatus": "FORBIDDEN"
            }
          }
        ]
      }
    },
    "404NaoEncontrado": {
      "description": "O recurso solicitado não foi encontrado",
      "schema": {
        "allOf": [
          {
            "$ref": "#/definitions/ApiError"
          },
          {
            "example": {
              "httpStatus": "NOT_FOUND"
            }
          }
        ]
      }
    },
...
    "500ErroInternoServidor": {
      "description": "Erro não identificado",
      "schema": {
        "allOf": [
          {
            "$ref": "#/definitions/ApiError"
          },
          {
            "example": {
              "httpStatus": "INTERNAL_SERVER_ERROR"
            }
          }
        ]
      }
    },
    "501NaoImplementado": {
      "description": "Não implementado",
      "schema": {
        "allOf": [
          {
            "$ref": "#/definitions/ApiError"
          },
          {
            "example": {
              "httpStatus": "NOT_IMPLEMENTED"
            }
          }
        ]
      }
    }
  },
  "definitions": {
    "ApiError": {
      "type": "object",
      "title": "Erro da requisição",
      "description": "Informações sobre o erro da requisição",
      "properties": {
        "codigo": {
          "type": "string",
          "description": "Código do erro"
        },
        "mensagem": {
          "type": "string",
          "description": "Mensagem do erro"
        },
        "dataHora": {
          "type": "string",
          "format": "date-time",
          "description": "Data e hora da ocorrência do erro"
        },
        "httpStatus": {
          "type": "string",
          "format": "HttpStatus",
          "description": "Código de status http do erro"
        },
        "erros": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "Lista de mensagens do erro"
        }
      },
      "required": [
        "codigo",
        "mensagem",
        "timestamp",
        "httpStatus"
      ],
      "example": {
        "codigo": "001A",
        "mensagem": "Ocorreu um erro.",
        "dataHora": "2023-03-28T22:35:10.123-03:00",
        "httpStatus": "INTERNAL_SERVER_ERROR",
        "erros": [
          "Detalhe 01 do erro.",
          "Detalhe 02 do erro."
        ]
      }
    },
...
    "Projeto": {
      "type": "object",
      "title": "Dados do projeto",
      "description": "Informações sobre o projeto",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": "1",
          "description": "Identificador do projeto",
          "readOnly": true
        },
        "nome": {
          "type": "string",
          "format": "string",
          "example": "Projeto 01",
          "description": "Nome do projeto"
        },
        "dataCriacao": {
          "type": "string",
          "format": "date",
          "example": "2020-01-01",
          "description": "Data de criação do projeto"
        },
        "funcionarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FuncionarioProjeto"
          },
          "description": "Funcionários envolvidos no projeto"
        }
      }
    },
    "FuncionarioProjeto": {
      "type": "object",
      "title": "Dados do funcionário",
      "description": "Informações sobre o funcionário",
      "properties": {
        "id": {
          "type": "number",
          "format": "long",
          "example": "1",
          "description": "Identificador do funcionário"
        },
        "nome": {
          "type": "string",
          "format": "string",
          "example": "Ricardo Carvalho Agostinho",
          "description": "Nome do funcionário"
        },
        "cpf": {
          "type": "string",
          "format": "string",
          "example": "12345678901",
          "pattern": "^[0-9]{11}$",
          "description": "O CPF do funcionário (apenas números)"
        },
        "email": {
          "type": "string",
          "format": "email",
          "example": "email@teste.com",
          "description": "E-mail do funcionário"
        },
        "salario": {
          "type": "number",
          "format": "BigDecimal",
          "example": "1000.00",
          "description": "Salário do funcionário"
        }
      }
    }
  },
...
}