# target/generated-sources/openapi/gen/orla/digital/desafiobackend/controller/v1/ProjetosApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package orla.digital.desafiobackend.controller.v1;

import orla.digital.desafiobackend.domain.dto.v1.ProjetoV1DTO;
import orla.digital.desafiobackend.domain.dto.v1.SearchFuncionario400ResponseV1DTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.HttpStatus;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-07-31T20:18:49.114273100-03:00[America/Fortaleza]")
@Validated
@RestController
@Tag(name = "Projeto", description = "Controles para projetos")
@RequestMapping("${openapi.desafioBackend.base-path:/desafio-backend/v1}")
public interface ProjetosApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /projetos : Criar projeto
     * Criar projeto
     *
     * @param novoProjeto Informações do novo projeto para inclusão. (required)
     * @return Created (status code 201)
     *         or Algum parâmetro informado está inválido (status code 400)
     *         or Não autorizado (status code 401)
     *         or Não permitido (status code 403)
     *         or Erro não identificado (status code 500)
     *         or Não implementado (status code 501)
     */
    @Operation(
        operationId = "createProjeto",
        summary = "Criar projeto",
        description = "Criar projeto",
        tags = { "Projeto" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created"),
            @ApiResponse(responseCode = "400", description = "Algum parâmetro informado está inválido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "401", description = "Não autorizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "403", description = "Não permitido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro não identificado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "501", description = "Não implementado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/projetos",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    @ResponseStatus(HttpStatus.CREATED)
    default void _createProjeto(
        @Parameter(name = "Novo Projeto", description = "Informações do novo projeto para inclusão.", required = true) @Valid @RequestBody ProjetoV1DTO novoProjeto
    ) {
        createProjeto(novoProjeto);
    }

    // Override this method
    default  void createProjeto(ProjetoV1DTO novoProjeto) {
        throw new IllegalArgumentException("Not implemented");

    }


    /**
     * GET /projetos : Buscar todos projetos cadastrados
     * Buscar todos os projetos cadastrados.
     *
     * @return Sucesso (status code 200)
     *         or Algum parâmetro informado está inválido (status code 400)
     *         or Não autorizado (status code 401)
     *         or Não permitido (status code 403)
     *         or Erro não identificado (status code 500)
     *         or Não implementado (status code 501)
     */
    @Operation(
        operationId = "searchProjeto",
        summary = "Buscar todos projetos cadastrados",
        description = "Buscar todos os projetos cadastrados.",
        tags = { "Projeto" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Sucesso", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProjetoV1DTO.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Algum parâmetro informado está inválido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "401", description = "Não autorizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "403", description = "Não permitido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro não identificado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "501", description = "Não implementado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/projetos",
        produces = { "application/json" }
    )
    @ResponseStatus(HttpStatus.OK)
    default List<ProjetoV1DTO> _searchProjeto(
        
    ) {
        return searchProjeto();
    }

    // Override this method
    default  List<ProjetoV1DTO> searchProjeto() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"nome\" : \"Projeto 01\", \"id\" : 1, \"funcionarios\" : [ { \"salario\" : 1000.0, \"cpf\" : \"12345678901\", \"nome\" : \"Ricardo Carvalho Agostinho\", \"id\" : 1.0, \"email\" : \"email@teste.com\" }, { \"salario\" : 1000.0, \"cpf\" : \"12345678901\", \"nome\" : \"Ricardo Carvalho Agostinho\", \"id\" : 1.0, \"email\" : \"email@teste.com\" } ], \"dataCriacao\" : \"2020-01-01T00:00:00.000+00:00\" }, { \"nome\" : \"Projeto 01\", \"id\" : 1, \"funcionarios\" : [ { \"salario\" : 1000.0, \"cpf\" : \"12345678901\", \"nome\" : \"Ricardo Carvalho Agostinho\", \"id\" : 1.0, \"email\" : \"email@teste.com\" }, { \"salario\" : 1000.0, \"cpf\" : \"12345678901\", \"nome\" : \"Ricardo Carvalho Agostinho\", \"id\" : 1.0, \"email\" : \"email@teste.com\" } ], \"dataCriacao\" : \"2020-01-01T00:00:00.000+00:00\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        throw new IllegalArgumentException("Not implemented");

    }


    /**
     * GET /projetos/{id} : Buscar projeto pelo identificador
     * Buscar projeto pelo identificador.
     *
     * @param id Id do projeto (required)
     * @return Sucesso (status code 200)
     *         or Algum parâmetro informado está inválido (status code 400)
     *         or Não autorizado (status code 401)
     *         or Não permitido (status code 403)
     *         or O recurso solicitado não foi encontrado (status code 404)
     *         or Erro não identificado (status code 500)
     *         or Não implementado (status code 501)
     */
    @Operation(
        operationId = "searchProjetoById",
        summary = "Buscar projeto pelo identificador",
        description = "Buscar projeto pelo identificador.",
        tags = { "Projeto" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Sucesso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProjetoV1DTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Algum parâmetro informado está inválido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "401", description = "Não autorizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "403", description = "Não permitido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "O recurso solicitado não foi encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro não identificado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "501", description = "Não implementado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/projetos/{id}",
        produces = { "application/json" }
    )
    @ResponseStatus(HttpStatus.OK)
    default ProjetoV1DTO _searchProjetoById(
        @Parameter(name = "id", description = "Id do projeto", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return searchProjetoById(id);
    }

    // Override this method
    default  ProjetoV1DTO searchProjetoById(Integer id) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"nome\" : \"Projeto 01\", \"id\" : 1, \"funcionarios\" : [ { \"salario\" : 1000.0, \"cpf\" : \"12345678901\", \"nome\" : \"Ricardo Carvalho Agostinho\", \"id\" : 1.0, \"email\" : \"email@teste.com\" }, { \"salario\" : 1000.0, \"cpf\" : \"12345678901\", \"nome\" : \"Ricardo Carvalho Agostinho\", \"id\" : 1.0, \"email\" : \"email@teste.com\" } ], \"dataCriacao\" : \"2020-01-01T00:00:00.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        throw new IllegalArgumentException("Not implemented");

    }


    /**
     * PUT /projetos/{id} : Atualizar projeto
     * Atualizar projeto
     *
     * @param id Id do projeto (required)
     * @param projeto Informações do projeto para atualização. (required)
     * @return Sucesso (status code 200)
     *         or Algum parâmetro informado está inválido (status code 400)
     *         or Não autorizado (status code 401)
     *         or Não permitido (status code 403)
     *         or O recurso solicitado não foi encontrado (status code 404)
     *         or Erro não identificado (status code 500)
     *         or Não implementado (status code 501)
     */
    @Operation(
        operationId = "updateProjeto",
        summary = "Atualizar projeto",
        description = "Atualizar projeto",
        tags = { "Projeto" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Sucesso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProjetoV1DTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Algum parâmetro informado está inválido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "401", description = "Não autorizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "403", description = "Não permitido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "O recurso solicitado não foi encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro não identificado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            }),
            @ApiResponse(responseCode = "501", description = "Não implementado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchFuncionario400ResponseV1DTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/projetos/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    @ResponseStatus(HttpStatus.OK)
    default ProjetoV1DTO _updateProjeto(
        @Parameter(name = "id", description = "Id do projeto", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "Projeto", description = "Informações do projeto para atualização.", required = true) @Valid @RequestBody ProjetoV1DTO projeto
    ) {
        return updateProjeto(id, projeto);
    }

    // Override this method
    default  ProjetoV1DTO updateProjeto(Integer id, ProjetoV1DTO projeto) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"nome\" : \"Projeto 01\", \"id\" : 1, \"funcionarios\" : [ { \"salario\" : 1000.0, \"cpf\" : \"12345678901\", \"nome\" : \"Ricardo Carvalho Agostinho\", \"id\" : 1.0, \"email\" : \"email@teste.com\" }, { \"salario\" : 1000.0, \"cpf\" : \"12345678901\", \"nome\" : \"Ricardo Carvalho Agostinho\", \"id\" : 1.0, \"email\" : \"email@teste.com\" } ], \"dataCriacao\" : \"2020-01-01T00:00:00.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        throw new IllegalArgumentException("Not implemented");

    }

}